requirements.txt
matplotlib==3.4.1
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.7
requests==2.25.1
telebot==0.0.4
